refactor(db): extract user CRUD to src/db/users.ts
Move AccessLevel constants, DbUser interface, and all user query functions
into their own module. Three functions that must invalidate the custom command
cache (upsertUser, updateTwitchBotEnabled, removeUser) are kept as thin wrappers
in db.ts so users.ts stays a pure DB layer with no knowledge of other caches.
All public exports are unchanged.
